items routes protected

+++ app/routes/item_routes.py
@@ -2,6 +2,7 @@ from flask import jsonify
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
 from sqlalchemy import desc
+from flask_jwt_extended import jwt_required, get_jwt_identity
 
 from app.models.item_model import Item
 from app.models.store_model import Store
@@ -66,29 +67,40 @@ class Items(MethodView):
 
     @item_blp.arguments(ItemCreateSchema)
     @item_blp.response(201, ItemCreateResponseSchema)
+    @jwt_required()
     def post(self, new_data):
         """Create Item"""
         try:
-            name = new_data.get("name")
-            store_id = new_data.get("store_id")
+            current_user = get_jwt_identity()
+            if current_user.get("role") == "admin":
+                name = new_data.get("name")
 
-            # check if store exists
-            store = Store.query.get(store_id)
-            if store is None:
-                return jsonify({"message": f"Store with id: {store_id} not found"}), 404
+                store_id = new_data.get("store_id")
 
-            # check for duplicate name of item
-            item = Item.query.filter_by(name=name).first()
+                # check if store exists
+                store = Store.query.get(store_id)
+                if store is None:
+                    return (
+                        jsonify({"message": f"Store with id: {store_id} not found"}),
+                        404,
+                    )
 
-            if item:
-                return jsonify({"message": "Item already exists"}), 400
+                # check for duplicate name of item
+                item = Item.query.filter_by(name=name).first()
+
+                if item:
+                    return jsonify({"message": "Item already exists"}), 400
+
+                else:
+                    item = Item(**new_data)
+                    db.session.add(item)
+                    db.session.commit()
+
+                    return item
 
             else:
-                item = Item(**new_data)
-                db.session.add(item)
-                db.session.commit()
+                return jsonify({"message": "Unauthorized access"}), 401
 
-                return item
         except Exception as e:
             print(e)
 
@@ -110,36 +122,56 @@ class ItemByID(MethodView):
 
     @item_blp.arguments(ItemSchema)
     @item_blp.response(200, ItemSchema)
+    @jwt_required()
     def patch(self, new_data, item_id):
         """Update Item By ID"""
         try:
-            item: Item = Item.query.get(item_id)
+            current_user = get_jwt_identity()
+            if current_user.get("role") == "admin":
+                item: Item = Item.query.get(item_id)
 
-            if item is None:
-                return jsonify({"message": f"Item with id: {item_id} not found"}), 404
+                if item is None:
+                    return (
+                        jsonify({"message": f"Item with id: {item_id} not found"}),
+                        404,
+                    )
 
-            item.name = new_data.get("name") or item.name
-            item.description = new_data.get("description") or item.description
-            item.unit_price = new_data.get("unit_price") or item.unit_price
-            item.store_id = new_data.get("store_id") or item.store_id
-            db.session.commit()
+                item.name = new_data.get("name") or item.name
+                item.description = new_data.get("description") or item.description
+                item.unit_price = new_data.get("unit_price") or item.unit_price
+                item.store_id = new_data.get("store_id") or item.store_id
+                db.session.commit()
 
-            return item
+                return item
+
+            else:
+                return jsonify({"message": "Unauthorized access"}), 401
+
         except Exception as e:
             print(e)
 
     @item_blp.response(204, ItemSchema)
+    @jwt_required()
     def delete(self, item_id):
         """Delete Item By ID"""
         try:
-            item = Item.query.get(item_id)
+            current_user = get_jwt_identity()
+            if current_user.get("role") == "admin":
+                item = Item.query.get(item_id)
 
-            if item is None:
-                return jsonify({"message": f"Item with id: {item_id} not found"}), 404
+                if item is None:
+                    return (
+                        jsonify({"message": f"Item with id: {item_id} not found"}),
+                        404,
+                    )
 
-            db.session.delete(item)
-            db.session.commit()
+                db.session.delete(item)
+                db.session.commit()
 
-            return item
+                return item
+
+            else:
+                return jsonify({"message": "Unauthorized access"}), 401
+
         except Exception as e:
             print(e)
